extract common class to cdc-base module
