+++ inlong-sort/sort-connectors/cdc-base/pom.xml
@@ -24,28 +24,22 @@
         <version>1.5.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>sort-connector-base</artifactId>
+    <artifactId>sort-connector-cdc-base</artifactId>
     <packaging>jar</packaging>
-    <name>Apache InLong - Sort-connector-base</name>
+    <name>Apache InLong - Sort-connector-cdc-base</name>
 
     <dependencies>
 
         <dependency>
             <groupId>org.apache.inlong</groupId>
-            <artifactId>audit-sdk</artifactId>
+            <artifactId>sort-connector-base</artifactId>
             <version>${project.version}</version>
         </dependency>
 
         <dependency>
-            <groupId>org.apache.inlong</groupId>
-            <artifactId>sort-format-json</artifactId>
-            <version>${project.version}</version>
+            <groupId>io.debezium</groupId>
+            <artifactId>debezium-embedded</artifactId>
             <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk-s3</artifactId>
         </dependency>
 
     </dependencies>
