+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/debezium/history/FlinkJsonTableChangeSerializer.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sort.base.debezium.history;
+package org.apache.inlong.sort.cdc.base.debezium.history;
 
 import io.debezium.document.Array;
 import io.debezium.document.Array.Entry;
