+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/debezium/internal/DebeziumOffset.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sort.base.debezium.internal;
+package org.apache.inlong.sort.cdc.base.debezium.internal;
 
 import java.io.Serializable;
 import java.util.Map;
