+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/util/DatabaseHistoryUtil.java
@@ -16,14 +16,14 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sort.base.util;
+package org.apache.inlong.sort.cdc.base.util;
 
 import io.debezium.relational.history.DatabaseHistory;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.inlong.sort.base.debezium.internal.SchemaRecord;
+import org.apache.inlong.sort.cdc.base.debezium.internal.SchemaRecord;
 
 /**
  * Util to safely visit schema history between {@link DatabaseHistory} and {@link DebeziumSourceFunction}.
