+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/util/TemporalConversions.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sort.base.util;
+package org.apache.inlong.sort.cdc.base.util;
 
 import java.time.Duration;
 import java.time.Instant;
