+++ inlong-sort/sort-connectors/oracle-cdc/src/main/java/org/apache/inlong/sort/cdc/oracle/OracleSource.java
@@ -23,8 +23,8 @@ import com.ververica.cdc.connectors.oracle.table.StartupOptions;
 import io.debezium.connector.oracle.OracleConnector;
 
 import java.util.Properties;
-import org.apache.inlong.sort.base.debezium.DebeziumDeserializationSchema;
-import org.apache.inlong.sort.base.debezium.internal.DebeziumOffset;
+import org.apache.inlong.sort.cdc.base.debezium.DebeziumDeserializationSchema;
+import org.apache.inlong.sort.cdc.base.debezium.internal.DebeziumOffset;
 import org.apache.inlong.sort.cdc.oracle.debezium.DebeziumSourceFunction;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
