+++ inlong-sort/sort-connectors/oracle-cdc/src/main/java/org/apache/inlong/sort/cdc/oracle/debezium/DebeziumSourceFunction.java
@@ -65,18 +65,18 @@ import org.apache.flink.streaming.api.functions.source.RichSourceFunction;
 import org.apache.flink.util.Collector;
 import org.apache.flink.util.ExceptionUtils;
 import org.apache.flink.util.FlinkRuntimeException;
-import org.apache.inlong.sort.base.debezium.DebeziumDeserializationSchema;
-import org.apache.inlong.sort.base.debezium.internal.DebeziumOffset;
-import org.apache.inlong.sort.base.debezium.internal.DebeziumOffsetSerializer;
-import org.apache.inlong.sort.base.debezium.internal.FlinkDatabaseHistory;
-import org.apache.inlong.sort.base.debezium.internal.FlinkOffsetBackingStore;
-import org.apache.inlong.sort.base.debezium.internal.Handover;
-import org.apache.inlong.sort.base.debezium.internal.SchemaRecord;
+import org.apache.inlong.sort.cdc.base.debezium.DebeziumDeserializationSchema;
+import org.apache.inlong.sort.cdc.base.debezium.internal.DebeziumOffset;
+import org.apache.inlong.sort.cdc.base.debezium.internal.DebeziumOffsetSerializer;
+import org.apache.inlong.sort.cdc.base.debezium.internal.FlinkDatabaseHistory;
+import org.apache.inlong.sort.cdc.base.debezium.internal.FlinkOffsetBackingStore;
+import org.apache.inlong.sort.cdc.base.debezium.internal.Handover;
+import org.apache.inlong.sort.cdc.base.debezium.internal.SchemaRecord;
 import org.apache.inlong.sort.base.metric.MetricOption;
 import org.apache.inlong.sort.base.metric.MetricOption.RegisteredMetric;
 import org.apache.inlong.sort.base.metric.MetricState;
 import org.apache.inlong.sort.base.metric.SourceMetricData;
-import org.apache.inlong.sort.base.util.DatabaseHistoryUtil;
+import org.apache.inlong.sort.cdc.base.util.DatabaseHistoryUtil;
 import org.apache.inlong.sort.base.util.MetricStateUtils;
 import org.apache.inlong.sort.cdc.oracle.debezium.internal.DebeziumChangeConsumer;
 import org.apache.inlong.sort.cdc.oracle.debezium.internal.DebeziumChangeFetcher;
