+++ inlong-sort/sort-connectors/oracle-cdc/src/main/java/org/apache/inlong/sort/cdc/oracle/debezium/internal/DebeziumChangeConsumer.java
@@ -25,8 +25,8 @@ import io.debezium.engine.DebeziumEngine.RecordCommitter;
 import java.util.List;
 import java.util.Map;
 import org.apache.flink.annotation.Internal;
-import org.apache.inlong.sort.base.debezium.internal.DebeziumOffset;
-import org.apache.inlong.sort.base.debezium.internal.Handover;
+import org.apache.inlong.sort.cdc.base.debezium.internal.DebeziumOffset;
+import org.apache.inlong.sort.cdc.base.debezium.internal.Handover;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.slf4j.Logger;
