+++ inlong-sort/sort-connectors/oracle-cdc/src/main/java/org/apache/inlong/sort/cdc/oracle/debezium/table/RowDataDebeziumDeserializeSchema.java
@@ -56,13 +56,13 @@ import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.table.types.logical.RowType;
 import org.apache.flink.types.RowKind;
 import org.apache.flink.util.Collector;
-import org.apache.inlong.sort.base.debezium.DebeziumDeserializationSchema;
-import org.apache.inlong.sort.base.debezium.table.AppendMetadataCollector;
-import org.apache.inlong.sort.base.debezium.table.DeserializationRuntimeConverter;
-import org.apache.inlong.sort.base.debezium.table.DeserializationRuntimeConverterFactory;
-import org.apache.inlong.sort.base.debezium.table.MetadataConverter;
-import org.apache.inlong.sort.base.util.RecordUtils;
-import org.apache.inlong.sort.base.util.TemporalConversions;
+import org.apache.inlong.sort.cdc.base.debezium.DebeziumDeserializationSchema;
+import org.apache.inlong.sort.cdc.base.debezium.table.AppendMetadataCollector;
+import org.apache.inlong.sort.cdc.base.debezium.table.DeserializationRuntimeConverter;
+import org.apache.inlong.sort.cdc.base.debezium.table.DeserializationRuntimeConverterFactory;
+import org.apache.inlong.sort.cdc.base.debezium.table.MetadataConverter;
+import org.apache.inlong.sort.cdc.base.util.RecordUtils;
+import org.apache.inlong.sort.cdc.base.util.TemporalConversions;
 import org.apache.kafka.connect.data.ConnectSchema;
 import org.apache.kafka.connect.data.Decimal;
 import org.apache.kafka.connect.data.Field;
