+++ inlong-sort/sort-connectors/pom.xml
@@ -34,6 +34,7 @@
 
     <modules>
         <module>base</module>
+        <module>cdc-base</module>
         <module>hive</module>
         <module>mysql-cdc</module>
         <module>kafka</module>
