+++ inlong-sort/sort-connectors/postgres-cdc/src/main/java/org.apache.inlong.sort.cdc.postgres/debezium/internal/DebeziumChangeFetcher.java
@@ -42,8 +42,8 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.streaming.api.functions.source.SourceFunction;
 import org.apache.flink.util.Collector;
-import org.apache.inlong.sort.base.debezium.DebeziumDeserializationSchema;
-import org.apache.inlong.sort.base.util.RecordUtils;
+import org.apache.inlong.sort.cdc.base.debezium.DebeziumDeserializationSchema;
+import org.apache.inlong.sort.cdc.base.util.RecordUtils;
 import org.apache.inlong.sort.cdc.postgres.connection.PostgreSQLJdbcConnectionOptions;
 import org.apache.inlong.sort.cdc.postgres.connection.PostgreSQLJdbcConnectionProvider;
 import org.apache.inlong.sort.cdc.postgres.manager.PostgreSQLQueryVisitor;
